refactor: JPA로 마이그레이션
