+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -2,7 +2,6 @@ package roomescape.dto;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationV2;
 
 public record ReservationResponse(
         long id,
@@ -12,12 +11,12 @@ public record ReservationResponse(
         ReservationThemeResponse theme) {
 
     public static ReservationResponse from(Reservation reservation) {
-        return new ReservationResponse(reservation.getId(), reservation.getName(),
+        return new ReservationResponse(reservation.getId(), reservation.getMember().getName(),
                 reservation.getDate(), ReservationTimeResponse.from(reservation.getTime()),
                 ReservationThemeResponse.from(reservation.getTheme()));
     }
 
-    public static ReservationResponse fromV2(ReservationV2 reservation) {
+    public static ReservationResponse fromV2(Reservation reservation) {
         return new ReservationResponse(reservation.getId(), reservation.getMember().getName(), reservation.getDate(),
                 ReservationTimeResponse.from(reservation.getTime()),
                 ReservationThemeResponse.from(reservation.getTheme()));
