+++ src/main/java/roomescape/repository/MemberRepository.java
@@ -6,8 +6,6 @@ import roomescape.domain.Member;
 
 public interface MemberRepository {
 
-    boolean existBySessionId(final String sessionId);
-
     boolean existByEmail(final String email);
 
     boolean existByName(final String name);
@@ -15,8 +13,6 @@ public interface MemberRepository {
     Member save(final Member member);
 
     Optional<Member> findByEmail(final String email);
-
-    void updateSessionId(final long memberId, final String sessionId);
 
     Optional<Member> findById(final long id);
