+++ src/main/java/roomescape/repository/jpa/ReservationJpaRepository.java
@@ -3,12 +3,16 @@ package roomescape.repository.jpa;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationV2;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 
 @Repository
-public interface ReservationJpaRepository extends JpaRepository<ReservationV2, Long> {
+public interface ReservationJpaRepository extends JpaRepository<Reservation, Long>, JpaSpecificationExecutor<Reservation> {
     boolean existsByDateAndTimeIdAndThemeId(LocalDate date, long timeId, long themeId);
 
-    List<ReservationV2> findByMemberId(Long memberId);
+    List<Reservation> findByMemberId(Long memberId);
+
+    boolean existsByDateAndTime(LocalDate date, ReservationTime time);
 }
