+++ src/main/resources/schema.sql
@@ -14,18 +14,6 @@ CREATE TABLE IF NOT EXISTS reservation_theme
     PRIMARY KEY (id)
 );
 
-CREATE TABLE IF NOT EXISTS reservation
-(
-    id   BIGINT       NOT NULL AUTO_INCREMENT,
-    name VARCHAR(255) NOT NULL,
-    date VARCHAR(255) NOT NULL,
-    time_id BIGINT,
-    theme_id BIGINT,                                        -- 컬럼 추가
-    PRIMARY KEY (id),
-    FOREIGN KEY (time_id) REFERENCES reservation_time (id),
-    FOREIGN KEY (theme_id) REFERENCES reservation_theme (id)            -- 외래키 추가
-);
-
 CREATE TABLE IF NOT EXISTS member
 (
     id   BIGINT       NOT NULL AUTO_INCREMENT,
@@ -37,7 +25,7 @@ CREATE TABLE IF NOT EXISTS member
     PRIMARY KEY (id)
 );
 
-CREATE TABLE IF NOT EXISTS reservation_v2
+CREATE TABLE IF NOT EXISTS reservation
 (
     id   BIGINT       NOT NULL AUTO_INCREMENT,
     member_id BIGINT,
