+++ src/test/java/roomescape/repository/ReservationRepositoryImplTest.java
@@ -1,0 +1,13 @@
+package roomescape.repository;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+
+class ReservationRepositoryImplTest {
+
+    @Test
+    void findByMemberIdAndThemeIdAndDateFromAndDateTo() {
+    }
+}
